feat: Remove support for PyTorch backend (#2611)
* Remove support for PyTorch from the codebase.
   - This is a breaking change.
   - Remove the PyTorch backend and optimize/opt_pytorch.py.
* Remove use of PyTorch from the tests.
   - Remove PyTorch lower bounds from tests/constraints.txt.
   - Remove comparison of tensor library backends in
     tests/test_backend_consistency.py as there is only NumPy and JAX now.
* Remove all docs and examples that mention or use PyTorch.
* Remove mentions of PyTorch in all parts of the website and REAMDE except
  for historical purposes like the release notes, JOSS paper, and notes
  about early pyhf research in FAQs.
   - For abstract use term "automatic differentiation".
* Update NumPy backend docstrings to accommodate NumPy v2.0 returns.

+++ src/pyhf/infer/calculators.py
@@ -121,7 +121,7 @@ class AsymptoticTestStatDistribution:
             >>> pyhf.set_backend("numpy")
             >>> bkg_dist = pyhf.infer.calculators.AsymptoticTestStatDistribution(0.0)
             >>> bkg_dist.cdf(0.0)
-            0.5
+            np.float64(0.5)
 
         Args:
             value (:obj:`float`): The test statistic value.
@@ -620,7 +620,7 @@ class EmpiricalDistribution:
             >>> samples = normal.sample((100,))
             >>> dist = pyhf.infer.calculators.EmpiricalDistribution(samples)
             >>> dist.expected_value(nsigma=1)
-            6.15094381...
+            np.float64(6.15094381...)
 
             >>> import pyhf
             >>> import numpy.random as random
